Changed room to remove lombok.

+++ src/main/java/com/dtg/meetroom/repository/RoomRepository.java
@@ -3,5 +3,5 @@ package com.dtg.meetroom.repository;
 import com.dtg.meetroom.model.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomRepository extends JpaRepository<Room, Long> {
+public interface RoomRepository extends JpaRepository<Room, Integer> {
 }
